Criada a função de haversine e feito alguns ajustes para usar builder e não o construtor

+++ src/main/java/br/com/involves/prospectIndicator/model/GeoLocatedObject.java
@@ -5,13 +5,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public abstract class GeoLocatedObject {
 
-    private BigDecimal longitude;
-    private BigDecimal latitude;
+    private Double longitude;
+
+    private Double latitude;
 }
